norm fixed echo and export
