+++ include/builtins.h
@@ -6,7 +6,7 @@
 /*   By: tanas <tanas@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/13 04:00:27 by sabdelra          #+#    #+#             */
-/*   Updated: 2023/08/23 20:17:33 by tanas            ###   ########.fr       */
+/*   Updated: 2023/08/24 10:08:12 by tanas            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -20,6 +20,7 @@ int		name_len(char *arg);
 int		_name(char *var1, char *var2);
 bool	check_var_export(char *arg, int *ret);
 bool	check_var_unset(char *arg);
+void	print_env_list(t_env **env_list);
 
 // BUILTINS
 int		ft_pwd(void);
