+++ src/builtins/ft_echo.c
@@ -21,32 +21,39 @@ static int	handle_tilde(char *arg, t_env *env_list)
 	return (0);
 }
 
-int		ft_echo(int argc, char **argv, t_env *env_list)
+static bool	process_flags(char **argv, int *i, bool *show_newline)
+{
+	int	j;
+
+	j = 1;
+	while (argv[*i][j] && argv[*i][j] == 'n')
+		j++;
+	if (argv[*i][j] && argv[*i][j] != 'n')
+	{
+		if (*show_newline)
+			*show_newline = true;
+		return (false);
+	}
+	else if (!argv[*i][j])
+	{
+		(*i)++;
+		*show_newline = false;
+		return (true);
+	}
+	(*i)++;
+	return (true);
+}
+
+int	ft_echo(int argc, char **argv, t_env *env_list)
 {
 	bool	show_newline;
 	int		i;
-	int		j;
 
 	show_newline = true;
 	i = 1;
 	while (i < argc && !ft_strncmp("-n", argv[i], 2))
-	{
-		j = 1;
-		while (argv[i][j] && argv[i][j] == 'n')
-			j++;
-		if (argv[i][j] && argv[i][j] != 'n')
-		{
-			show_newline = true;
+		if (!process_flags(argv, &i, &show_newline))
 			break ;
-		}
-		else if (!argv[i][j])
-		{
-			i++;
-			show_newline = false;
-			continue ;
-		}
-		i++;
-	}
 	while (i < argc)
 	{
 		if (!handle_tilde(argv[i], env_list))
